requirements.txt
aiosmtplib==1.1.6
anyio==3.6.2
asgiref==3.6.0
attrs==22.2.0
Automat==22.10.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
constantly==15.1.0
cryptography==40.0.1
cssselect==1.2.0
Deprecated==1.2.13
dnspython==2.3.0
fastapi==0.95.0
filelock==3.10.7
geographiclib==2.0
geopy==2.3.0
googlemaps==4.10.0
h11==0.14.0
httpcore==0.16.3
httptools==0.5.0
httpx==0.23.3
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
itemadapter==0.8.0
itemloaders==1.0.6
jmespath==1.0.1
lxml==4.9.2
packaging==23.0
parsel==1.7.0
phonenumbers==8.12.48
Pillow==9.5.0
pip==22.3.1
Protego==0.2.1
pyasn1-modules==0.2.8
pyasn1==0.4.8
pycparser==2.21
pycryptodome==3.10.4
pydantic==1.10.7
PyDispatcher==2.0.7
PyJWT==2.6.0
pymongo==4.3.3
pyOpenSSL==23.1.1
python-dotenv==1.0.0
pytz==2023.3
PyYAML==6.0
queuelib==1.6.2
requests-file==1.5.1
requests==2.28.2
rfc3986==1.5.0
Scrapy==2.8.0
service-identity==21.1.0
setuptools==65.5.1
six==1.16.0
sniffio==1.3.0
starlette==0.26.1
supertokens-python==0.12.5
tldextract==3.1.0
twilio==7.9.1
twisted-iocpsupport==1.0.2
Twisted==22.10.0
typing-extensions==4.5.0
urllib3==1.26.15
uvicorn==0.21.1
w3lib==2.1.1
watchfiles==0.19.0
websockets==10.4
wheel==0.38.4
wrapt==1.15.0
zope.interface==6.0
